--- OctoAwesome/OctoAwesome.Basics/Blocks/StoneBrickBlockDefinition.cs
@@ -1,45 +1,0 @@
-﻿using System;
-using System.Drawing;
-
-namespace OctoAwesome.Basics
-{
-    public sealed class StoneBrickBlockDefinition : BlockDefinition
-    {
-        public override string Name
-        {
-            get { return Languages.OctoBasics.StoneBrick; }
-        }
-
-        public override Bitmap Icon
-        {
-            get { return (Bitmap)Bitmap.FromFile("./Assets/OctoAwesome.Basics/Blocks/brick_grey.png"); }
-        }
-
-
-        public override Bitmap[] Textures
-        {
-            get
-            {
-                return new[] {
-                    (Bitmap)Bitmap.FromFile("./Assets/OctoAwesome.Basics/Blocks/brick_grey.png"),
-                };
-            }
-        }
-
-        public override PhysicalProperties GetProperties(ILocalChunkCache manager, int x, int y, int z)
-        {
-            return new PhysicalProperties()
-            {
-                Density = 2.5f,
-                FractureToughness = 0.1f,
-                Granularity = 0.1f,
-                Hardness = 0.9f
-            };
-        }
-
-        public override void Hit(IBlockDefinition block, PhysicalProperties itemProperties)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
