+++ OctoAwesome/OctoAwesome.Basics/Definitions/Blocks/SandBlockDefinition.cs
@@ -1,0 +1,46 @@
+﻿using System;
+using System.Drawing;
+
+namespace OctoAwesome.Basics
+{
+    public sealed class SandBlockDefinition : BlockDefinition
+    {
+        public override string Name
+        {
+            get { return Languages.OctoBasics.RedCotton; }
+        }
+
+        public override Bitmap Icon
+        {
+            get { return (Bitmap)Bitmap.FromFile("./Assets/OctoAwesome.Basics/Blocks/sand.png"); }
+        }
+
+
+        public override Bitmap[] Textures
+        {
+            get
+            {
+                return new[] {
+                    (Bitmap)Bitmap.FromFile("./Assets/OctoAwesome.Basics/Blocks/sand.png")
+                };
+            }
+        }
+
+        public override void Hit(IBlockDefinition block, PhysicalProperties itemProperties)
+        {
+            throw new NotImplementedException();
+        }
+
+        public override PhysicalProperties GetProperties(ILocalChunkCache manager, int x, int y, int z)
+        {
+            return new PhysicalProperties()
+            {
+                //Schüttdichte
+                Density = 1.5f,
+                FractureToughness = 0.3f,
+                Granularity = 0.9f,
+                Hardness = 0.1f
+            };
+        }
+    }
+}
