+++ OctoAwesome/OctoAwesome.Basics/Definitions/Items/PickAxeDefinition.cs
