+++ OctoAwesome/OctoAwesome.Basics/Definitions/Items/Pickaxe.cs
@@ -1,0 +1,9 @@
+﻿namespace OctoAwesome.Basics
+{
+    public class Pickaxe : Item
+    {
+        public override void Hit(IItem item)
+        {
+        }
+    }
+}
