--- OctoAwesome/OctoAwesome.Basics/Items/Pickaxe.cs
@@ -1,9 +1,0 @@
-﻿namespace OctoAwesome.Basics
-{
-    public class Pickaxe : Item
-    {
-        public override void Hit(IItem item)
-        {
-        }
-    }
-}
