+++ OctoAwesome/OctoAwesome.Basics/Languages/OctoBasics.de.resx
@@ -117,6 +117,9 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
+  <data name="BirchWood" xml:space="preserve">
+    <value>Birkenholz</value>
+  </data>
   <data name="BlueCotton" xml:space="preserve">
     <value>Blauer Stoff</value>
   </data>
@@ -140,6 +143,9 @@
   </data>
   <data name="Leaves" xml:space="preserve">
     <value>Blätter</value>
+  </data>
+  <data name="OrangeLeaves" xml:space="preserve">
+    <value>Orange Blätter</value>
   </data>
   <data name="RedCotton" xml:space="preserve">
     <value>Roter Stoff</value>
