+++ OctoAwesome/OctoAwesome.Basics/Languages/OctoBasics.resx
@@ -117,6 +117,9 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
+  <data name="BirchWood" xml:space="preserve">
+    <value>Birch Wood</value>
+  </data>
   <data name="BlueCotton" xml:space="preserve">
     <value>Blue Cotton</value>
   </data>
@@ -140,6 +143,9 @@
   </data>
   <data name="Leaves" xml:space="preserve">
     <value>Leaves</value>
+  </data>
+  <data name="OrangeLeaves" xml:space="preserve">
+    <value>Orange Leaves</value>
   </data>
   <data name="RedCotton" xml:space="preserve">
     <value>Red Cotton</value>
